Add user protecting on endpoints

+++ app/app.py
@@ -2,7 +2,7 @@ import uuid
 from typing import Optional
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.schemas import CreatePost, PostResponse, UserCreate, UserUpdate, UserRead
-from app.db import Post, create_db_and_tables, get_async_session
+from app.db import Post, create_db_and_tables, get_async_session, User
 from app.images import imagekit
 from app.users import auth_backend, current_active_user, fastapi_users
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -38,6 +38,7 @@ app.include_router(fastapi_users.get_users_router(UserRead, UserUpdate), prefix=
 async def upload_file(
         file: UploadFile = File(...),
         caption: str = Form(...),
+        user: User = Depends(current_active_user),
         session: AsyncSession = Depends(get_async_session)
 ):
     # File extension type/name
@@ -65,6 +66,7 @@ async def upload_file(
         )
 
         post = Post(
+            user_id=user.id,
             caption=caption,
             url=response.url,
             file_type="video" if file.content_type.startswith("video/") else "image",
@@ -96,7 +98,8 @@ async def upload_file(
 
 @app.get("/feed")
 async def get_feed(
-        session: AsyncSession = Depends(get_async_session)
+        session: AsyncSession = Depends(get_async_session),
+        user: User = Depends(current_active_user)
 ):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     all_posts = result.scalars().all()
@@ -106,11 +109,13 @@ async def get_feed(
     for post in all_posts:
         post = {
             "id": str(post.id),
+            "user_id": str(post.user_id),
             "caption": post.caption,
             "url": post.url,
             "file_type": post.file_type,
             "file_name": post.file_name,
             "created_at": post.created_at.isoformat(),
+            "is_owner": post.user_id == user.id
         }
 
         post_data.append(post)
@@ -120,7 +125,11 @@ async def get_feed(
 
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(
+        post_id: str,
+        session: AsyncSession = Depends(get_async_session),
+        user: User = Depends(current_active_user)
+):
 
     try:
         post_uuid = uuid.UUID(post_id)
@@ -130,6 +139,9 @@ async def delete_post(post_id: str, session: AsyncSession = Depends(get_async_se
 
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
+
+        if post.user_id != user.id:
+            raise HTTPException(status_code=404, detail="You don't have a permission to delete this post")
 
         await session.delete(post)
         await session.commit()
